# swaggerDocs.json
{
  "swagger": "2.0",
  "info": {
    "description": "FoodFactory API information",
    "version": "1.0.0",
    "title": "Food Factory"
  },
  "host": "localhost:8000",
  "basePath": "/",
...
  "paths": {
    "/signup": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "signup page",
        "description": "to open signup page (here just sends successful response)",
        "operationId": "singupForm",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "successful operation"
          }
        }
      },
      "post": {
        "tags": [
          "users"
        ],
        "summary": "User registeration",
        "description": "User data for registeration",
        "operationId": "userSignUp",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "reqBody",
            "description": "User signup credentials",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "email": {
                  "type": "string"
                },
                "role": {
                  "type": "string",
                  "description": "Specify whether admin or user role"
                },
                "password": {
                  "type": "string"
                }
              },
              "required": [
                "name",
                "email",
                "role",
                "password"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfull registeration"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Authentication failed"
          }
        }
      }
    },
    "/login": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "login page",
        "description": "sends response to open login page",
        "operationId": "loginForm",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "successful operation"
          }
        }
      },
      "post": {
        "tags": [
          "users"
        ],
        "summary": "User login",
        "description": "Submit user credentials to login",
        "operationId": "userSignIn",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "reqBody",
            "description": "User login credentials",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfull login"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Authentication failed"
          }
        }
      }
    },
    "/logout":{
      "post" : {
        "tags": [
          "users"
        ],
        "summary": "User logout",
        "description": "Clears login credentials from express session - logout only if user logged in",
        "operationId": "userLogout",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
        ],
        "responses": {
          "200": {
            "description": "Successfull Logout"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Authentication failed"
          }
        }
      }
    },
    "/updatePassword": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "updatePassword page",
        "description": "to open updatePassword page (here just sends success response)",
        "operationId": "updatePasswordForm",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "successful response"
          }
        }
      },
      "post": {
        "tags": [
          "users"
        ],
        "summary": "update Password",
        "description": "User can change their passowrd by entering their registered email and password",
        "operationId": "updatePassword",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "reqBody",
            "description": "Update Password credentials",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string"
                },
                "oldPassword": {
                  "type": "string",
                  "description": "Enter here the old password"
                },
                "password": {
                  "type": "string",
                  "description": "Enter new password"
                },
                "confirmPassword": {
                  "type": "string",
                  "description": "Enter new password again"
                }
              },
              "required": [
                "oldPassword",
                "email",
                "password",
                "confirmPassword"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Password updated successfully"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Authentication failed"
          }
        }
      }
    },
    "/updateUser": {
      "put": {
        "tags": [
          "users"
        ],
        "summary": "update user name",
        "description": "User can change name - logged in user is choosen",
        "operationId": "updateUser",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "reqBody",
            "description": "Update logged in user name",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                }
              },
              "required": [
                "name"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User name updated successfully"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Authentication failed"
          }
        }
      }
    },
    "/listAllUsers": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "list all users",
        "description": "List of all registered users. This can be accessed only by users with admin role",
        "operationId": "listAllUsers",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "successful response"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Authentication failed"
          }
        }
      }
    },
    "/deactivateUser/{userId}": {
      "put": {
        "tags": [
          "users"
        ],
        "summary": "Deactivate user",
        "description": "Admin can deactivate user by passing id and changing status to false",
        "operationId": "deactivateUser",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "description": "user id that need to be deactivated"
          }
        ],
        "responses": {
          "200": {
            "description": "User deactivated successfully"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Authentication failed"
          }
        }
      }
    },
...
  },
...
}